Maya plugin refactor

Maya plugin now uses the model lib to compile the model as well as saving out a raw model when compiling.

As a result the maya plugin now links to engines core and creates a job system and filesystem so that the model lib used by the plugin has a complete engine core to work with.

See merge request !27

+++ Pulp/Core/DiskFile.h
@@ -20,12 +20,12 @@ public:
 		return file_.valid();
 	}
 
-	virtual uint32_t read(void* buffer, uint32_t length) X_OVERRIDE;
-	virtual uint32_t write(const void* buffer, uint32_t length) X_OVERRIDE;
-	virtual void seek(size_t position, SeekMode::Enum origin) X_OVERRIDE;
-	virtual size_t remainingBytes(void) const X_OVERRIDE;
-	virtual size_t tell(void) const X_OVERRIDE;
-	virtual void setSize(size_t numBytes) X_OVERRIDE;
+	virtual size_t read(void* buffer, size_t length) X_OVERRIDE;
+	virtual size_t write(const void* buffer, size_t length) X_OVERRIDE;
+	virtual void seek(int64_t position, SeekMode::Enum origin) X_OVERRIDE;
+	virtual uint64_t remainingBytes(void) const X_OVERRIDE;
+	virtual uint64_t tell(void) const X_OVERRIDE;
+	virtual void setSize(int64_t numBytes) X_OVERRIDE;
 
 private:
