+++ Pulp/RenderSystems/Common/Textures/TextureLoaderJPG.cpp
@@ -48,11 +48,11 @@ namespace JPG
 		{
 			jpeg_xfile_src_mgr* src = reinterpret_cast<jpeg_xfile_src_mgr*>(cinfo->src);
 
-			uint32 bytes_read;
+			size_t bytes_read;
 			bytes_read = src->file->read(src->buffer, JPEG_MGR_BUFFER_SIZE);
 
 			src->mgr.next_input_byte = src->buffer;
-			src->mgr.bytes_in_buffer = (size_t)bytes_read;
+			src->mgr.bytes_in_buffer = bytes_read;
 			if (0 == src->mgr.bytes_in_buffer)
 			{
 				/* The image file is truncated. We insert EOI marker to tell the library to stop processing. */
@@ -249,7 +249,7 @@ namespace JPG
 		jpeg_destroy_decompress(&cinfo);
 
 #if X_DEBUG == 1
-		size_t left = file->remainingBytes();
+		uint64_t left = file->remainingBytes();
 		X_WARNING_IF(left > 0, "TextureJPG", "potential read fail, bytes left in file: %i", left);
 #endif
