+++ Pulp/RenderSystems/Common/Textures/TextureLoaderPSD.cpp
@@ -350,7 +350,7 @@ namespace PSD
 
 
 #if X_DEBUG == 1
-		size_t left = file->remainingBytes();
+		uint64_t left = file->remainingBytes();
 		X_WARNING_IF(left > 0, "TexturePSD", "potential read fail, bytes left in file: %i", left);
 #endif
