+++ Pulp/RenderSystems/Common/Textures/TextureLoaderTGA.cpp
@@ -139,7 +139,7 @@ XTextureFile* XTexLoaderTGA::loadTexture(core::XFile* file)
 
 	// read the image data.
 
-	uint32_t bytes_read = file->read(img->pFaces[0], DataSize);
+	size_t bytes_read = file->read(img->pFaces[0], DataSize);
 
 	if (bytes_read != DataSize)
 	{
@@ -151,7 +151,7 @@ XTextureFile* XTexLoaderTGA::loadTexture(core::XFile* file)
 	}
 
 #if X_DEBUG == 1
-	size_t left = file->remainingBytes();
+	uint64_t left = file->remainingBytes();
 	X_WARNING_IF(left > 0, "TextureTGA", "potential read fail, bytes left in file: %i", left);
 #endif
